Move _iotools to top level so that it can be used by compas.json_* functions
