+++ src/compas/_iotools.py
