+++ src/compas/files/_xml/xml_pre_38.py
@@ -3,7 +3,7 @@ from __future__ import division
 from __future__ import print_function
 
 import xml.etree.ElementTree as ET
-from compas.files import _iotools
+from compas import _iotools
 from compas.files._xml.xml_cpython import prettify_string  # doesn't need special handling for pre-3.8 so we just import
 
 __all__ = [
