+++ tests/compas/fixtures/iotools/image.png
